[DROOLS-6519] More code cleanup
