+++ drools-model/drools-mvel-compiler/src/test/java/org/drools/mvel/EvaluatorGenerator.java
@@ -58,7 +58,8 @@ public class EvaluatorGenerator {
 
         defineLastStatement(compiledMVELBlock);
 
-        logGenerateClass();
+        // disabled to prevent long loggings, reactivate when debugging
+        // logGenerateClass();
         return template;
     }
