--- drools-model/drools-mvel-compiler/src/test/java/org/drools/mvel/OptimizerFactory.java
@@ -1,9 +1,0 @@
-package org.drools.mvel;
-
-public class
-OptimizerFactory {
-
-    public static void setDefaultOptimizer(String reflective) {
-
-    }
-}
