--- drools-model/drools-mvel-compiler/src/test/java/org/drools/mvel/ParserConfiguration.java
@@ -1,9 +1,0 @@
-package org.drools.mvel;
-
-public class ParserConfiguration {
-
-
-    public void addImport(String math, Class<?> mathClass) {
-
-    }
-}
