--- drools-model/drools-mvel-compiler/src/test/java/org/drools/mvel/SampleBeanAccessor.java
@@ -1,12 +1,0 @@
-package org.drools.mvel;
-
-public class SampleBeanAccessor {
-  public Object getProperty(String name, Object contextObj, Object variableFactory) {
-    return contextObj;
-  }
-
-  public Object setProperty(String name, Object contextObj, Object variableFactory, Object value) {
-    return contextObj;
-  }
-
-}
