+++ drools-model/drools-mvel-compiler/src/test/java/org/drools/mvel/TestInterface.java
@@ -1,12 +1,5 @@
 package org.drools.mvel;
 
-/**
- * Created by IntelliJ IDEA.
- * User: brockm
- * Date: Jun 22, 2007
- * Time: 6:31:04 PM
- * To change this template use File | Settings | File Templates.
- */
 public interface TestInterface {
   public String getName();
